This is the implementation idea for exposing our query endpoint
group.js: I had to make it let cuz distribution object was not available yet. But it will be available when the function is called.
Node.js: We have a startup function already. During the startup phase, we have to put the crawler group. We also have additional work to do. Like running the actual crawler and running the indexer.

Query.js: The new endpoint that will be exposed. We can call this endpoint like awsIP:awsPORT/query/get with the query passed as an object. This will execute the query subsystem.

ADDITIONAL WORK:
1. We have to think about how we will spawn the rest of the nodes.
2. We also have to define serialization and deserializion libraries in the frontend.
3. We have to run the crawler and indexer workflow during startup.
4. We can also have a hashmap to return the values since our function is idempotent.

LET ME KNOW IF THIS SEEMS LIKE A GOOD IMPLEMENTATION OR IF ANYTHING NEEDS TO BE CHANGED